+++ lib/rubocop/cop/style/character_literal.rb
@@ -32,13 +32,11 @@ module RuboCop
 
         # Dummy implementation of method in ConfigurableEnforcedStyle that is
         # called from StringHelp.
-        def opposite_style_detected
-        end
+        def opposite_style_detected; end
 
         # Dummy implementation of method in ConfigurableEnforcedStyle that is
         # called from StringHelp.
-        def correct_style_detected
-        end
+        def correct_style_detected; end
       end
     end
   end